Factor out ScreenWithPredictions
