+++ browser/src/neovim/index.ts
@@ -2,6 +2,7 @@ export * from "./actions"
 export * from "./EventContext"
 export * from "./MsgPack"
 export * from "./Screen"
+export * from "./ScreenWithPredictions"
 export * from "./Session"
 export * from "./NeovimProcessSpawner"
 export * from "./NeovimInstance"
